requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
attrs==24.2.0
dataclasses==0.6
frozenlist==1.5.0
# Editable install with no version control (glue==0.1.0)
-e /Users/cooper/Desktop/AI_ML/Creating/ParadiseLabs/My AI Projects/glue_framework/glue
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==75.5.0
typing_extensions==4.12.2
yarl==1.17.1
